python-client/requirements.txt
cryptography>=42.0.0
argon2-cffi>=23.1.0
zeroconf>=0.131.0
pytest>=8.0.0
